Make the complete Windows TZ tests an author test.
When trying to install DateTime::TimeZone, this can happen.

0. An older version of DateTime::TimeZone is already installed.
1. cpanm sees DateTime::TimeZone::Local::Win32 as a dependency.
2. cpanm tests DateTime::TimeZone::Local::Win32.
3. DateTime::TimeZone::Local::Win32 sees DateTime::TimeZone::Local is installed and runs its tests.
4. The tests fail because the installed DateTime::TimeZone is out of date.

This failure is usually because the installed DateTime::TimeZone doesn't have
every time zone in the Windows TZ database. This isn't a critical problem
for the user.

Only have the release tests make sure the basics work.  Leave checking
the whole TZ database for an author test.